Fix #473 - messages only rendered on whole seconds and not fractions of seconds

+++ TwitchDownloaderCore/ChatRenderer.cs
@@ -336,7 +336,7 @@ namespace TwitchDownloaderCore
         {
             List<CommentSection> commentList = new List<CommentSection>();
             SKBitmap newFrame = new SKBitmap(renderOptions.ChatWidth, renderOptions.ChatHeight);
-            double currentTimeSeconds = currentTick / renderOptions.Framerate;
+            double currentTimeSeconds = currentTick / (double)renderOptions.Framerate;
             int newestCommentIndex = chatRoot.comments.FindLastIndex(x => x.content_offset_seconds <= currentTimeSeconds);
 
             if (newestCommentIndex == lastestUpdate?.CommentIndex)
